// src/services/types/admin-permission-group-bulk-update-request.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export type AdminPermissionGroupBulkUpdateRequest = {
	/**
	 * @type integer, int64
	 */
	permissionGroupId: number;
	/**
	 * @type array
	 */
	adminIds: number[];
};