// src/services/types/admin-summary-response.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export const adminSummaryResponseProviderEnum = {GOOGLE: "GOOGLE", APPLE: "APPLE"} as const;

export type AdminSummaryResponseProviderEnumKey = (typeof adminSummaryResponseProviderEnum)[keyof typeof adminSummaryResponseProviderEnum];

export type AdminSummaryResponse = {
	/**
	 * @description 어드민 아이디
	 * @type integer, int64
	 */
	id: number;
	/**
	 * @description 권한
	 * @type string
	 */
	permissionGroupName: string;
	/**
	 * @description 소셜 유형
	 * @type string | undefined
	 */
	provider?: AdminSummaryResponseProviderEnumKey;
	/**
	 * @description 로그인 아이디
	 * @type string
	 */
	loginId: string;
	/**
	 * @description 이름
	 * @type string
	 */
	name: string;
	/**
	 * @description 삭제 여부
	 * @type boolean
	 */
	removed: boolean;
	/**
	 * @description 잠금 일시
	 * @type string | undefined, date-time
	 */
	lockedAt?: string;
	/**
	 * @description 마지막 로그인 일시
	 * @type string | undefined, date-time
	 */
	lastLoginAt?: string;
	/**
	 * @description 생성 일시
	 * @type string | undefined, date-time
	 */
	createdAt?: string;
};